requirements.txt
schwab-py>=1.0.0
pandas>=2.0.0
pandas_ta>=0.3.14b
numpy>=1.24.0
yfinance>=0.2.28
pytz>=2023.3
psycopg2-binary>=2.9.9
